refactor: announcements to tsx
