+++ components/announcements/amount.tsx
