+++ components/announcements/hcb_cta.tsx
@@ -1,36 +1,40 @@
 import { Box, Button, Grid, Heading, Text } from 'theme-ui'
 import Icon from '@hackclub/icons'
 import NextLink from 'next/link'
-import { thousands } from '../../lib/members'
 
-export default function SlackCTA() {
+export default function HCBCTA() {
   return (
     <Box
       as="section"
       sx={{
         bg: 'orange',
-        backgroundImage: t => t.util.gx('yellow', 'orange'),
+        backgroundImage: (t: any) => t.util.gx('yellow', 'orange'),
         color: 'white',
         py: [4, 5]
       }}
     >
       <Grid gap={[3, 4]} columns={[null, 'auto 1fr']} variant="layout.copy">
-        <Icon glyph="welcome" size={72} />
+        <Icon glyph="bank-account" size={72} />
         <Box>
           <Heading as="h2" variant="headline" mt={0}>
-            Teenager? New here? Welcome!
+            Looking to start a nonprofit?
           </Heading>
           <Text variant="subtitle" sx={{ lineHeight: 'caption', mb: 3 }}>
-            Hack Club is a global community of high school makers & student-led
-            coding clubs. We’ve got a 24/7 Slack chatroom of {thousands}k+
-            teenagers learning to code & building amazing projects, & you’ll fit
-            right in.
+            We're accepting applications! No startup fees, no minimum balance, and no long wait time.
           </Text>
           <br />
           <br />
-          <NextLink href="/" passHref>
+          <NextLink href="/fiscal-sponsorship" passHref>
             <Button bg="cyan" as="a">
               Learn more
+            </Button>
+          </NextLink>
+
+          &nbsp;&nbsp;&nbsp;
+
+          <NextLink href="https://nonprofit.new" passHref>
+            <Button bg="orange" as="a">
+              Apply now
             </Button>
           </NextLink>
         </Box>
