+++ components/announcements/holder.tsx
